// packages/storefront/src/app/common/generated-types.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};

export type CustomField = {
  description?: Maybe<Array<LocalizedString>>;
  internal?: Maybe<Scalars['Boolean']>;
  label?: Maybe<Array<LocalizedString>>;
  list: Scalars['Boolean'];
  name: Scalars['String'];
  readonly?: Maybe<Scalars['Boolean']>;
  type: Scalars['String'];
};

export type ErrorResult = {
  errorCode: ErrorCode;
  message: Scalars['String'];
};

export type Node = {
  id: Scalars['ID'];
};
// ...
export type Channel = Node & {
  __typename?: 'Channel';
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars['JSON']>;
  defaultLanguageCode: LanguageCode;
  defaultShippingZone?: Maybe<Zone>;
  defaultTaxZone?: Maybe<Zone>;
  id: Scalars['ID'];
  pricesIncludeTax: Scalars['Boolean'];
  token: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};

export type Collection = Node & {
  __typename?: 'Collection';
  assets: Array<Asset>;
  breadcrumbs: Array<CollectionBreadcrumb>;
  children?: Maybe<Array<Collection>>;
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  description: Scalars['String'];
  featuredAsset?: Maybe<Asset>;
  filters: Array<ConfigurableOperation>;
  id: Scalars['ID'];
  languageCode?: Maybe<LanguageCode>;
  name: Scalars['String'];
  parent?: Maybe<Collection>;
  position: Scalars['Int'];
  productVariants: ProductVariantList;
  slug: Scalars['String'];
  translations: Array<CollectionTranslation>;
  updatedAt: Scalars['DateTime'];
};

export type CollectionProductVariantsArgs = {
  options?: Maybe<ProductVariantListOptions>;
};

export type CollectionBreadcrumb = {
  __typename?: 'CollectionBreadcrumb';
  id: Scalars['ID'];
  name: Scalars['String'];
  slug: Scalars['String'];
};

export type CollectionList = PaginatedList & {
  __typename?: 'CollectionList';
  items: Array<Collection>;
  totalItems: Scalars['Int'];
};
// ...
export type Country = Node & {
  __typename?: 'Country';
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  enabled: Scalars['Boolean'];
  id: Scalars['ID'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  translations: Array<CountryTranslation>;
  updatedAt: Scalars['DateTime'];
};
// ...
export type CurrentUser = {
  __typename?: 'CurrentUser';
  channels: Array<CurrentUserChannel>;
  id: Scalars['ID'];
  identifier: Scalars['String'];
};

export type CurrentUserChannel = {
  __typename?: 'CurrentUserChannel';
  code: Scalars['String'];
  id: Scalars['ID'];
  permissions: Array<Permission>;
  token: Scalars['String'];
};

export type Customer = Node & {
  __typename?: 'Customer';
  addresses?: Maybe<Array<Address>>;
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  emailAddress: Scalars['String'];
  firstName: Scalars['String'];
  id: Scalars['ID'];
  lastName: Scalars['String'];
  orders: OrderList;
  phoneNumber?: Maybe<Scalars['String']>;
  title?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
  user?: Maybe<User>;
};
// ...
export type Facet = Node & {
  __typename?: 'Facet';
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  id: Scalars['ID'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  translations: Array<FacetTranslation>;
  updatedAt: Scalars['DateTime'];
  values: Array<FacetValue>;
};

export type FacetList = PaginatedList & {
  __typename?: 'FacetList';
  items: Array<Facet>;
  totalItems: Scalars['Int'];
};

export type FacetTranslation = {
  __typename?: 'FacetTranslation';
  createdAt: Scalars['DateTime'];
  id: Scalars['ID'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};

export type FacetValue = Node & {
  __typename?: 'FacetValue';
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  facet: Facet;
  id: Scalars['ID'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  translations: Array<FacetValueTranslation>;
  updatedAt: Scalars['DateTime'];
};
// ...
export type OrderList = PaginatedList & {
  __typename?: 'OrderList';
  items: Array<Order>;
  totalItems: Scalars['Int'];
};
// ...
export type PaymentMethodQuote = {
  __typename?: 'PaymentMethodQuote';
  code: Scalars['String'];
  eligibilityMessage?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  isEligible: Scalars['Boolean'];
  customFields?: Maybe<Scalars['JSON']>;
};

/** The price range where the result has more than one price */
export type PriceRange = {
  __typename?: 'PriceRange';
  max: Scalars['Int'];
  min: Scalars['Int'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** The active Channel */
  activeChannel: Channel;
  /** The active Customer */
  activeCustomer?: Maybe<Customer>;
  /**
   * The active Order. Will be `null` until an Order is created via `addItemToOrder`. Once an Order reaches the
   * state of `PaymentApproved` or `PaymentSettled`, then that Order is no longer considered "active" and this
   * query will once again return `null`.
   */
  activeOrder?: Maybe<Order>;
  /** An array of supported Countries */
  availableCountries: Array<Country>;
  /** Returns a Collection either by its id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
  collection?: Maybe<Collection>;
  /** A list of Collections available to the shop */
  collections: CollectionList;
  /** Returns a list of payment methods and their eligibility based on the current active Order */
  eligiblePaymentMethods: Array<PaymentMethodQuote>;
  /** Returns a list of eligible shipping methods based on the current active Order */
  eligibleShippingMethods: Array<ShippingMethodQuote>;
  /** Returns information about the current authenticated User */
  me?: Maybe<CurrentUser>;
  /** Returns the possible next states that the activeOrder can transition to */
  nextOrderStates: Array<Scalars['String']>;
  /**
   * Returns an Order based on the id. Note that in the Shop API, only orders belonging to the
   * currently-authenticated User may be queried.
   */
  order?: Maybe<Order>;
  /**
   * Returns an Order based on the order `code`. For guest Orders (i.e. Orders placed by non-authenticated Customers)
   * this query will only return the Order within 2 hours of the Order being placed. This allows an Order confirmation
   * screen to be shown immediately after completion of a guest checkout, yet prevents security risks of allowing
   * general anonymous access to Order data.
   */
  orderByCode?: Maybe<Order>;
  /** Get a Product either by id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
  product?: Maybe<Product>;
  /** Get a list of Products */
  products: ProductList;
  /** Search Products based on the criteria set by the `SearchInput` */
  search: SearchResponse;
};
// ...
export enum SortOrder {
  ASC = 'ASC',
  DESC = 'DESC',
}
// ...
export type SearchResultSortParameter = {
  name?: Maybe<SortOrder>;
  price?: Maybe<SortOrder>;
};
// ...
type DiscriminateUnion<T, U> = T extends U ? T : never;
// ...
  export type Variables = UpdateAddressMutationVariables;
// ...
  export type Items = NonNullable<
    NonNullable<NonNullable<NonNullable<GetOrderListQuery['activeCustomer']>['orders']>['items']>[number]
  >;
// ...
  export type FeaturedAsset = NonNullable<
    NonNullable<
      NonNullable<NonNullable<GetCollectionsQuery['collections']>['items']>[number]
    >['featuredAsset']
  >;
// ...
  export type FocalPoint = NonNullable<AssetFragment['focalPoint']>;
// ...
export namespace Country {
  export type Fragment = CountryFragment;
}
// ...
  export type Breadcrumbs = NonNullable<
    NonNullable<
      NonNullable<
        NonNullable<NonNullable<GetProductDetailQuery['product']>['collections']>[number]
      >['breadcrumbs']
    >[number]
  >;
// ...
export namespace SearchProducts {
  export type Variables = SearchProductsQueryVariables;
  export type Query = SearchProductsQuery;
  export type Search = NonNullable<SearchProductsQuery['search']>;
  export type Items = NonNullable<NonNullable<NonNullable<SearchProductsQuery['search']>['items']>[number]>;
  export type PriceWithTax = NonNullable<
    NonNullable<NonNullable<NonNullable<SearchProductsQuery['search']>['items']>[number]>['priceWithTax']
  >;
  export type PriceRangeInlineFragment = DiscriminateUnion<
    NonNullable<
      NonNullable<NonNullable<NonNullable<SearchProductsQuery['search']>['items']>[number]>['priceWithTax']
    >,
    { __typename?: 'PriceRange' }
  >;
  export type ProductAsset = NonNullable<
    NonNullable<NonNullable<NonNullable<SearchProductsQuery['search']>['items']>[number]>['productAsset']
  >;
  export type FocalPoint = NonNullable<
    NonNullable<
      NonNullable<NonNullable<NonNullable<SearchProductsQuery['search']>['items']>[number]>['productAsset']
    >['focalPoint']
  >;
  export type FacetValues = NonNullable<
    NonNullable<NonNullable<SearchProductsQuery['search']>['facetValues']>[number]
  >;
  export type FacetValue = NonNullable<
    NonNullable<NonNullable<NonNullable<SearchProductsQuery['search']>['facetValues']>[number]>['facetValue']
  >;
  export type Facet = NonNullable<
    NonNullable<
      NonNullable<
        NonNullable<NonNullable<SearchProductsQuery['search']>['facetValues']>[number]
      >['facetValue']
    >['facet']
  >;
}

export namespace GetCollection {
  export type Variables = GetCollectionQueryVariables;
  export type Query = GetCollectionQuery;
  export type Collection = NonNullable<GetCollectionQuery['collection']>;
  export type FeaturedAsset = NonNullable<NonNullable<GetCollectionQuery['collection']>['featuredAsset']>;
  export type Breadcrumbs = NonNullable<
    NonNullable<NonNullable<GetCollectionQuery['collection']>['breadcrumbs']>[number]
  >;
  export type Children = NonNullable<
    NonNullable<NonNullable<GetCollectionQuery['collection']>['children']>[number]
  >;
  export type _FeaturedAsset = NonNullable<
    NonNullable<
      NonNullable<NonNullable<GetCollectionQuery['collection']>['children']>[number]
    >['featuredAsset']
  >;
}
// ...
export type UpdateAddressMutationVariables = Exact<{
  input: UpdateAddressInput;
}>;
// ...
export type SearchProductsQueryVariables = Exact<{
  input: SearchInput;
}>;

export type SearchProductsQuery = { __typename?: 'Query' } & {
  search: { __typename?: 'SearchResponse' } & Pick<SearchResponse, 'totalItems'> & {
      items: Array<
        { __typename?: 'SearchResult' } & Pick<
          SearchResult,
          'productId' | 'slug' | 'productName' | 'description' | 'productVariantId' | 'inStock'
        > & {
            priceWithTax:
              | ({ __typename?: 'PriceRange' } & Pick<PriceRange, 'min' | 'max'>)
              | { __typename?: 'SinglePrice' };
            productAsset?: Maybe<
              { __typename?: 'SearchResultAsset' } & Pick<SearchResultAsset, 'id' | 'preview'> & {
                  focalPoint?: Maybe<{ __typename?: 'Coordinate' } & Pick<Coordinate, 'x' | 'y'>>;
                }
            >;
            facetValueIds?: Maybe<Array<Scalars['ID']>>;
          }
      >;
      facetValues: Array<
        { __typename?: 'FacetValueResult' } & Pick<FacetValueResult, 'count'> & {
            facetValue: { __typename?: 'FacetValue' } & Pick<FacetValue, 'id' | 'name' | 'code'> & {
                facet: { __typename?: 'Facet' } & Pick<Facet, 'id' | 'name' | 'code'>;
              };
          }
      >;
    };
};

export type GetCollectionQueryVariables = Exact<{
  id?: Maybe<Scalars['ID']>;
  slug?: Maybe<Scalars['String']>;
}>;

export type GetCollectionQuery = { __typename?: 'Query' } & {
  collection?: Maybe<
    { __typename?: 'Collection' } & Pick<Collection, 'id' | 'name' | 'slug' | 'description'> & {
        featuredAsset?: Maybe<{ __typename?: 'Asset' } & AssetFragment>;
        breadcrumbs: Array<
          { __typename?: 'CollectionBreadcrumb' } & Pick<CollectionBreadcrumb, 'id' | 'slug' | 'name'>
        >;
        children?: Maybe<
          Array<
            { __typename?: 'Collection' } & Pick<Collection, 'id' | 'slug' | 'name'> & {
                featuredAsset?: Maybe<{ __typename?: 'Asset' } & AssetFragment>;
              }
          >
        >;
      }
  >;
};
// ...
export type OrderGroup = Node & {
  __typename?: 'OrderGroup';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  alias: Scalars['String'];
  code: Scalars['String'];
  state: Scalars['String'];
  active: Scalars['Boolean'];
  customer?: Customer;
  channel?: Channel;
  orders?: OrderList;
  ordersQuantity: Scalars['Int'];
  totalWithTax: Scalars['Int'];
};